+++ website/docusaurus.config.js
@@ -133,7 +133,7 @@ module.exports = {
         },
         announcementBar: {
             id: 'announcement-bar-',
-            content: `Learn about the latest beta release of Crawlee for Python from the creators of Crawlee. Join us live at 9 AM EST on August 5th, 2024. <a target="_blank" rel="noopener noreferrer" href="https://apify.com/resources/crawlee-for-python-webinar">Sign up now!</a> ⭐`,
+            content: `🎉️ <b>Learn about Crawlee for Python from its creators. Join us at 9 am EST on August 5.<a href="https://apify.com/resources/crawlee-for-python-webinar">Sign up now!</a></b> 🥳️`,
         },
         navbar: {
             hideOnScroll: true,