+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.0](https://github.com/apify/crawlee-python/releases/tag/v0.2.0) (2024-08-05)
+
+### 🚀 Features
+
+- Add new curl impersonate HTTP client ([#387](https://github.com/apify/crawlee-python/issues/387)) ([9c06260](https://github.com/apify/crawlee-python/commit/9c06260c0ee958522caa9322001a3186e9e43af4))
+- *(playwright)* `infinite_scroll` helper ([#393](https://github.com/apify/crawlee-python/issues/393)) ([34f74bd](https://github.com/apify/crawlee-python/commit/34f74bdcffb42a6c876a856e1c89923d9b3e60bd))
+
 ## [0.1.2](https://github.com/apify/crawlee-python/releases/tag/v0.1.2) (2024-07-30)
 
 ### 🚀 Features